// types/interfaces/ICallEntry.ts
export interface ICallEntry {
  Event: String;
  Privilege: string;
  Channel: string;
  ChannelState: string;
  ChannelStateDesc: string;
  CallerIDNum: string;
  CallerIDName: string;
  ConnectedLineNum: string;
  ConnectedLineName: string;
  Language: string;
  AccountCode: string;
  Context: string;
  Exten: string;
  Priority: string;
  Uniqueid: string;
  Linkedid: string;
  Cause: string;
  "Cause-txt": string;
}